Added confirms to delete user and card.

+++ src/web/client/js/app/controllers/admin.js
@@ -85,7 +85,7 @@ hgotsAdmin.controller('AdminController', [ '$scope', '$routeParams', '$window', 
   });
 }]);
 
-hgotsAdmin.controller('AdminUserController', [ '$scope', '$routeParams', '$http', '$location', 'AdminShared', 'User', 'Card', function($scope, $routeParams, $http, $location, AdminShared, User, Card) {
+hgotsAdmin.controller('AdminUserController', [ '$scope', '$routeParams', '$http', '$location', 'AdminShared', 'User', 'Card', 'confirm', function($scope, $routeParams, $http, $location, AdminShared, User, Card, confirm) {
   // set all user types
   $scope.userTypes = AdminShared.userTypes;
   // watch for user changes
@@ -171,8 +171,12 @@ hgotsAdmin.controller('AdminUserController', [ '$scope', '$routeParams', '$http'
   };
   
   $scope.deleteUser = function() {
-    $scope.user.$delete(function() {
-      AdminShared.showUser(null);
+    confirm('Do you really want to delete ' + $scope.user.username + '?').then(function(confirm) {
+      if ( confirm ) {
+        $scope.user.$delete(function() {
+          AdminShared.showUser(null);
+        });
+      }
     });
   };
   
@@ -206,10 +210,14 @@ hgotsAdmin.controller('AdminUserController', [ '$scope', '$routeParams', '$http'
     });
   };
   $scope.deleteCard = function(card) {
-    $http({ method: 'DELETE', url: '/api/v2/users/' + $scope.user.id + '/cards/' + card.id }).success(function() {
-      $scope.user.cards = $scope.user.cards.filter(function(cardItem) {
-        return ( cardItem.id !== card.id );
-      });
+    confirm("Do you really want to remove card '" + card.uid + "' from this user?").then(function(confirm) {
+      if ( confirm ) {
+        $http({ method: 'DELETE', url: '/api/v2/users/' + $scope.user.id + '/cards/' + card.id }).success(function() {
+          $scope.user.cards = $scope.user.cards.filter(function(cardItem) {
+            return ( cardItem.id !== card.id );
+          });
+        });
+      }
     });
   };
 }]);
